requirements.txt
# Community Mesh Platform
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Meshtastic
meshtastic>=2.3.0
pubsub>=0.1.2

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
greenlet
paho-mqtt>=2.0.0

# LLM
httpx>=0.26.0
openai>=1.12.0

# Templates
jinja2>=3.1.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
pyserial>=3.5

# Knowledge Base (Kiwix + optional full RAG)
beautifulsoup4>=4.12.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
datasets>=2.14.0
tqdm>=4.66.0
